Move DockerfileWriter & ServiceYAMLWriter in packs package
